+++ configuration.php
@@ -8,7 +8,7 @@
  * • include all php classes
  *
  * @author      David Voglgsang
- * @version     1.0
+ * @version     1.1
  *
 */
 
@@ -65,20 +65,18 @@ Table of Contents:
 
   /* 1.3 REGISTER ALL THE CLASSES
   /------------------------*/
-  // register mastertheme classes
   function prefix_registerFunction($class) {
     if ( 0 !== strpos( $class, 'prefix_' ) ) {
         return;
     }
-    require_once( get_template_directory() . "/classes/$class/class.$class.php" );
-  }
-  // register childtheme classes
-  function prefix_registerFunction_Child($class) {
-    if ( 0 !== strpos( $class, 'prefix_' ) ) {
-        return;
-    }
-    if(prefix_core_BaseFunctions::CheckDir(get_stylesheet_directory() . '/classes')):
-      require_once( get_stylesheet_directory() . "/classes/$class/class.$class.php" );
+    $url = get_template_directory() . "/classes/$class/class.$class.php";
+    $url_child = get_stylesheet_directory() . "/classes/$class/class.$class.php";
+    if(file_exists($url)):
+      // register mastertheme classes
+      require_once($url);
+    elseif(file_exists($url_child)):
+      // register childtheme classes
+      require_once($url_child);
     endif;
   }
 
@@ -132,7 +130,6 @@ Table of Contents:
     // register
     add_action( 'init', function(){
       spl_autoload_register('prefix_registerFunction');
-      spl_autoload_register('prefix_registerFunction_Child');
     }, 1 );
     // register
     add_action( 'init', 'prefix_RunClassesInit', 1 );
@@ -141,7 +138,6 @@ Table of Contents:
     prefix_JSON_as_Global();
     // register
     spl_autoload_register('prefix_registerFunction');
-    spl_autoload_register('prefix_registerFunction_Child');
     // run classes init
     prefix_RunClassesInit();
   endif;